🚀 [Feat]: query hook service에서 분리 및 useMyStudyRole 추가
